+++ app/src/main/java/com/example/aquariuminformation/ui/commonui/CommonUI.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -13,7 +14,8 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Star
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -25,8 +27,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
@@ -35,11 +39,12 @@ import com.example.aquariuminformation.R
 import com.example.aquariuminformation.ui.theme.AquariumInformationTheme
 import com.example.aquariuminformation.ui.theme.Shapes
 
+
 @Composable
 fun TitleWideCard(
 	modifier: Modifier = Modifier,
 	title: String,
-	icon: ImageVector,
+	icon: Painter,
 	content: @Composable ColumnScope.() -> Unit,
 ){
 	Column(
@@ -51,7 +56,7 @@ fun TitleWideCard(
 			verticalAlignment = Alignment.CenterVertically
 		) {
 			Icon(
-				imageVector = icon,
+				painter = icon,
 				contentDescription = title
 			)
 			HeaderTextLarge(
@@ -70,64 +75,145 @@ fun TitleWideCard(
 
 @Composable
 fun IconTextRow(
-	icon: ImageVector,
+	modifier: Modifier = Modifier,
+	icon: Painter,
 	iconTint: Color = MaterialTheme.colorScheme.onBackground,
 	text: String,
 	textColor: Color = MaterialTheme.colorScheme.onBackground
 ){
-	Row(
-		modifier = Modifier
-			.fillMaxWidth()
-			.padding(dimensionResource(id = R.dimen.padding_verySmall)),
-		verticalAlignment = Alignment.CenterVertically
-	) {
-		Icon(
+	Column(modifier = modifier) {
+		Row(
 			modifier = Modifier
-				.weight(1f),
-			imageVector =icon,
-			contentDescription = null,
-			tint = iconTint
-		)
-		Text(
-			modifier = Modifier
-				.weight(5f),
-			text = text,
-			color = textColor
-		)
+				.fillMaxWidth()
+				.padding(dimensionResource(id = R.dimen.padding_verySmall)),
+			verticalAlignment = Alignment.CenterVertically
+		) {
+			Icon(
+				modifier = Modifier
+					.weight(1f),
+				painter = icon,
+				contentDescription = null,
+				tint = iconTint
+			)
+			Text(
+				modifier = Modifier
+					.weight(5f),
+				text = text,
+				color = textColor
+			)
+		}
 	}
 }
 
 @Composable
 fun TextRow(
+	modifier: Modifier = Modifier,
 	text1: String,
 	text2: String,
-
 	){
-	Row(
-		modifier = Modifier
-			.fillMaxWidth()
-			.padding(dimensionResource(id = R.dimen.padding_verySmall)),
-		verticalAlignment = Alignment.CenterVertically
-	) {
-		Text(
+	Column(modifier = modifier) {
+		Row(
 			modifier = Modifier
-				.weight(1f),
-			text = text1
+				.fillMaxWidth()
+				.padding(dimensionResource(id = R.dimen.padding_verySmall)),
+			verticalAlignment = Alignment.CenterVertically
+		) {
+			Text(
+				modifier = Modifier
+					.weight(1f),
+				text = text1
+			)
+			Text(
+				modifier = Modifier
+					.weight(1f),
+				text = text2
+			)
+		}
+	}
+
+}
+
+@Composable
+fun NavButton(
+	modifier: Modifier = Modifier,
+	shape: Shape = Shapes.large,
+	containerColor: Color = MaterialTheme.colorScheme.background,
+	contentColor: Color = MaterialTheme.colorScheme.onBackground,
+	title: String,
+	icon: ImageVector
+){
+	Button(
+		modifier = modifier,
+		onClick = { /*TODO*/ },
+		shape = shape,
+		colors = ButtonDefaults.buttonColors(
+			containerColor = containerColor,
+			contentColor = contentColor,
+		),
+		elevation = ButtonDefaults.buttonElevation(
+			defaultElevation = dimensionResource(id = R.dimen.elevation_large),
+			pressedElevation = dimensionResource(id = R.dimen.elevation_small)
 		)
-		Text(
+	) {
+		Column(
 			modifier = Modifier
-				.weight(1f),
-			text = text2
+				.padding(dimensionResource(id = R.dimen.padding_small)),
+			verticalArrangement = Arrangement.Center,
+			horizontalAlignment = Alignment.CenterHorizontally,
+		) {
+			Text(
+				text = title
+			)
+			Icon(
+				imageVector = icon,
+				contentDescription = null
+			)
+		}
+
+	}
+}
+
+@Composable
+fun NavButtonRow(
+	modifier: Modifier = Modifier,
+	title1: String,
+	icon1: ImageVector,
+	title2: String,
+	icon2: ImageVector,
+	containerColor: Color = MaterialTheme.colorScheme.background,
+	contentColor: Color = MaterialTheme.colorScheme.onBackground,
+){
+	Row(
+		modifier = modifier
+	) {
+		NavButton(
+			modifier = Modifier
+				.weight(1f)
+				.padding(dimensionResource(id = R.dimen.padding_small)),
+			title = title1,
+			icon = icon1,
+			containerColor = containerColor,
+			contentColor = contentColor,
+		)
+		NavButton(
+			modifier = Modifier
+				.weight(1f)
+				.padding(dimensionResource(id = R.dimen.padding_small)),
+			title = title2,
+			icon = icon2,
+			containerColor = containerColor,
+			contentColor = contentColor,
 		)
 	}
 }
 
 @Composable
-fun FishComCard(){
-	SingleWideCard(
-		backgroundCardColor = MaterialTheme.colorScheme.tertiary,
-		contentColor = MaterialTheme.colorScheme.onTertiary
-	) {
+fun FishComCard(modifier: Modifier = Modifier){
+	Column(modifier = modifier) {
+		SingleWideCard(
+			backgroundCardColor = MaterialTheme.colorScheme.tertiary,
+			contentColor = MaterialTheme.colorScheme.onTertiary
+		) {
 
 			Row(
 				modifier = Modifier.fillMaxWidth(),
@@ -137,7 +223,7 @@ fun FishComCard(){
 				Icon(
 					modifier = Modifier
 						.weight(1f),
-					imageVector = Icons.Filled.Star, 
+					painter = painterResource(id = R.drawable.baseline_new_releases_24),
 					contentDescription = null)
 				Column(
 					modifier = Modifier
@@ -152,15 +238,64 @@ fun FishComCard(){
 					)
 				}
 			}
+		}
 	}
 }
+
+//@Composable
+//fun AppInfo(
+//) {
+//
+//	Row(
+//		horizontalArrangement = Arrangement.Center,
+//	) {
+//		Text(
+//			text = stringResource(R.string.app_version)
+//		)
+//		Text(
+//			text = stringResource(id = R.string.text_label_version, version),
+//		)
+//	}
+//}
 
 @Composable
 fun PageView(
 	content: @Composable ColumnScope.() -> Unit,
 ){
 	Column(
-		modifier = Modifier.verticalScroll(rememberScrollState()),
+		modifier = Modifier
+			.fillMaxSize()
+			.verticalScroll(rememberScrollState()),
+		horizontalAlignment = Alignment.CenterHorizontally
+	) {
+		Spacer(modifier = Modifier.height(dimensionResource(id = R.dimen.padding_small)))
+		content()
+	}
+}
+
+@Composable
+fun PageViewCenter(
+	content: @Composable ColumnScope.() -> Unit,
+){
+	Column(
+		modifier = Modifier
+			.fillMaxSize()
+			.verticalScroll(rememberScrollState()),
+		horizontalAlignment = Alignment.CenterHorizontally,
+		verticalArrangement = Arrangement.Center,
+	) {
+		Spacer(modifier = Modifier.height(dimensionResource(id = R.dimen.padding_small)))
+		content()
+	}
+}
+
+@Composable
+fun PageViewLazy(
+	content: @Composable ColumnScope.() -> Unit,
+){
+	Column(
+		modifier = Modifier
+			.fillMaxSize(),
 		horizontalAlignment = Alignment.CenterHorizontally
 	) {
 		Spacer(modifier = Modifier.height(dimensionResource(id = R.dimen.padding_small)))
@@ -170,81 +305,93 @@ fun PageView(
 
 @Composable
 fun SingleWideCard(
+	modifier: Modifier = Modifier,
 	shape: Shape = Shapes.large,
 	backgroundCardColor: Color = MaterialTheme.colorScheme.background,
 	contentColor: Color = MaterialTheme.colorScheme.onBackground,
 	content: @Composable ColumnScope.() -> Unit,
 ){
-	ElevatedCard(
-		modifier = Modifier
-			.padding(
-				top = dimensionResource(id = R.dimen.padding_small),
-				bottom = dimensionResource(id = R.dimen.padding_small),
+	Column(modifier = modifier) {
+		ElevatedCard(
+			modifier = Modifier
+				.padding(
+					top = dimensionResource(id = R.dimen.padding_small),
+					bottom = dimensionResource(id = R.dimen.padding_small),
 //				start = dimensionResource(id = R.dimen.padding_medium),
 //				end = dimensionResource(id = R.dimen.padding_medium)
+				)
+				.fillMaxWidth(fraction = 0.9f),
+			shape = shape,
+			colors = CardDefaults.cardColors(
+				containerColor = backgroundCardColor,
+				contentColor = contentColor
+			),
+			elevation = CardDefaults.cardElevation(
+				defaultElevation = dimensionResource(id = R.dimen.elevation_medium)
 			)
-			.fillMaxWidth(fraction = 0.9f),
-		shape = shape,
-		colors = CardDefaults.cardColors(
-			containerColor = backgroundCardColor,
-			contentColor = contentColor
-		),
-		elevation = CardDefaults.cardElevation(
-			defaultElevation = dimensionResource(id = R.dimen.elevation_medium)
-		)
-	) {
-		Column(
-			modifier = Modifier
-				.padding(dimensionResource(id = R.dimen.padding_small)),
 		) {
-			content()
+			Column(
+				modifier = Modifier
+					.padding(dimensionResource(id = R.dimen.padding_small)),
+			) {
+				content()
+			}
 		}
 	}
 }
 
 @Composable
 fun HeaderTextCard(
+	modifier: Modifier = Modifier,
 	text: String,
 	style: TextStyle = MaterialTheme.typography.titleMedium,
 	textAlign: TextAlign = TextAlign.Center
 ){
-	Text(
-		modifier = Modifier.padding(start = dimensionResource(id = R.dimen.padding_small)),
-		text = text,
-		style = style,
-		textAlign = textAlign,
-	)
+	Column(modifier = modifier) {
+		Text(
+			modifier = Modifier.padding(start = dimensionResource(id = R.dimen.padding_small)),
+			text = text,
+			style = style,
+			textAlign = textAlign,
+		)
+	}
 }
 
 @Composable
 fun BodyTextCard(
+	modifier: Modifier = Modifier,
 	text: String,
 	style: TextStyle = MaterialTheme.typography.bodyMedium,
 	textAlign: TextAlign = TextAlign.Center
 ){
-	Text(
-		modifier = Modifier.padding(dimensionResource(id = R.dimen.padding_small)),
-		text = text,
-		style = style,
-		textAlign = textAlign,
-	)
+	Column(modifier = modifier) {
+		Text(
+			modifier = Modifier.padding(dimensionResource(id = R.dimen.padding_small)),
+			text = text,
+			style = style,
+			textAlign = textAlign,
+		)
+	}
 }
 
 @Composable
 fun HeaderTextLarge(
+	modifier: Modifier = Modifier,
 	text: String,
 	color: Color = MaterialTheme.colorScheme.onBackground,
 	style: TextStyle = MaterialTheme.typography.titleLarge,
 	textAlign: TextAlign = TextAlign.Start
 ){
-	Text(
-		modifier = Modifier
-			.padding(start = dimensionResource(id = R.dimen.padding_small)),
-		text = text,
-		color = color,
-		style = style,
-		textAlign = textAlign,
-	)
+	Column(modifier = modifier) {
+		Text(
+			modifier = Modifier
+				.padding(start = dimensionResource(id = R.dimen.padding_small)),
+			text = text,
+			color = color,
+			style = style,
+			textAlign = textAlign,
+		)
+	}
 }
 
 //@Composable
@@ -301,6 +448,41 @@ fun HeaderTextLarge(
 @ExperimentalMaterial3Api
 @Preview(showBackground = true)
 @Composable
+fun NavButtonPreview(){
+	AquariumInformationTheme {
+		Column(
+			modifier = Modifier
+				.background(color = MaterialTheme.colorScheme.background)
+		){
+			NavButton(
+				title = stringResource(R.string.converters),
+				icon = Icons.Filled.Home
+			)
+		}
+	}
+}
+
+@ExperimentalMaterial3Api
+@Preview(showBackground = true)
+@Composable
+fun NavButtonPreviewDark(
+){
+	AquariumInformationTheme(useDarkTheme = true) {
+		Column(
+			modifier = Modifier
+				.background(color = MaterialTheme.colorScheme.background)
+		){
+			NavButton(
+				title = stringResource(R.string.converters),
+				icon = Icons.Filled.Home
+			)
+		}
+	}
+}
+
+@ExperimentalMaterial3Api
+@Preview(showBackground = true)
+@Composable
 fun FisComPreview(){
 	AquariumInformationTheme {
 		Column(
@@ -338,7 +520,7 @@ fun CardTitlePreview(){
 		){
 			TitleWideCard(
 				title = stringResource(id = R.string.app_name),
-				icon = Icons.Filled.Home
+				icon = painterResource(id = R.drawable.baseline_info_24)
 			) {
 				SingleWideCard {
 					BodyTextCard(text = stringResource(id = R.string.text_welcome))
@@ -360,7 +542,7 @@ fun CardTitlePreviewDark(
 		){
 			TitleWideCard(
 				title = stringResource(id = R.string.app_name),
-				icon = Icons.Filled.Home
+				icon = painterResource(id = R.drawable.baseline_info_24)
 			){
 				SingleWideCard {
 					BodyTextCard(text = stringResource(id = R.string.text_welcome))