+++ app/src/main/java/com/example/aquariuminformation/ui/pages/InfoPage.kt
@@ -3,22 +3,17 @@ package com.example.aquariuminformation.ui.pages
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccountCircle
-import androidx.compose.material.icons.filled.Build
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.aquariuminformation.R
 import com.example.aquariuminformation.ui.commonui.BodyTextCard
+import com.example.aquariuminformation.ui.commonui.HeaderTextCard
 import com.example.aquariuminformation.ui.commonui.IconTextRow
 import com.example.aquariuminformation.ui.commonui.PageView
 import com.example.aquariuminformation.ui.commonui.SingleWideCard
@@ -26,30 +21,13 @@ import com.example.aquariuminformation.ui.commonui.TextRow
 import com.example.aquariuminformation.ui.commonui.TitleWideCard
 import com.example.aquariuminformation.ui.theme.AquariumInformationTheme
 
+
 @Composable
 fun InfoPage(){
 	PageView {
 		InfoLayout()
 	}
 }
-
-//@Composable
-//fun InfoCards(
-//	modifier: Modifier = Modifier
-//){
-//	Column(modifier = modifier) {
-//		LazyColumn(
-//			content = {
-//				items(infoData) {
-//					InfoItem(
-//						dataSource = it,
-//						modifier = Modifier.padding(dimensionResource(R.dimen.padding_small)),
-//					)
-//				}
-//			}
-//		)
-//	}
-//}
 
 @Composable
 fun InfoLayout(
@@ -61,7 +39,7 @@ fun InfoLayout(
 	) {
 		TitleWideCard(
 			title = stringResource(id = R.string.text_title_info),
-			icon = Icons.Filled.Info
+			icon = painterResource(id = R.drawable.baseline_info_24)
 		){
 			SingleWideCard {
 				BodyTextCard(text = stringResource(id = R.string.text_info_1))
@@ -71,7 +49,7 @@ fun InfoLayout(
 		}
 		TitleWideCard(
 			title = stringResource(R.string.errors_or_bugs),
-			icon = Icons.Filled.Build
+			icon = painterResource(id = R.drawable.baseline_build_24)
 		) {
 			SingleWideCard(
 				backgroundCardColor = MaterialTheme.colorScheme.primary,
@@ -82,27 +60,28 @@ fun InfoLayout(
 		}
 		TitleWideCard(
 			title = stringResource(R.string.contact),
-			icon = Icons.Filled.Search
+			icon = painterResource(id = R.drawable.baseline_person_search_24)
 		){
 			SingleWideCard {
 				IconTextRow(
-					icon = Icons.Filled.Email,
+					icon = painterResource(id = R.drawable.baseline_email_24),
 					text = stringResource(id = R.string.text_email)
 				)
 				IconTextRow(
-					icon = Icons.Filled.AccountCircle,
+					icon = painterResource(id = R.drawable.baseline_web_24),
 					text = stringResource(id = R.string.text_website)
 				)
 			}
 		}
 		TitleWideCard(
 			title = stringResource(R.string.app_information),
-			icon = Icons.Filled.Settings
+			icon = painterResource(id = R.drawable.baseline_settings_suggest_24),
 		) {
 			SingleWideCard {
+				HeaderTextCard(text = stringResource(id = R.string.app_name))
 				TextRow(
-					text1 = stringResource(id = R.string.text_label_version),
-					text2 = "2.0"
+					text1 = stringResource(id = R.string.app_version),
+					text2 = stringResource(id = R.string.appVersion)
 				)
 			}
 		}