+++ app/src/main/assets/Changelog.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 * Ability to remove ADs with in*app billing
 * Marine tank types added to Water Hardness
 * Gesture Navigation
+* Launch Default Page
 
 ## [Issues]